+++ lbh15/properties/bismuth_thermochemical_properties/bismuth_thermochemical.py
@@ -5,10 +5,10 @@ from typing import Union
 import numpy as np
 from scipy.constants import atm
 from scipy.constants import R
-from ..interface import PropertyInterface
 from ..common_interface import OxygenPartialPressureInterface
 from ..common_interface import MolarEnthalpyInterface
 from ..common_interface import MolarEntropyInterface
+from ..common_interface import GibbsFreeEnergyInterface
 from ..bismuth_properties import h
 from ..._commons import BISMUTH_BOILING_TEMPERATURE as T_b0
 from ..._commons import BISMUTH_MELTING_TEMPERATURE as T_m0
@@ -351,7 +351,7 @@ class MolarEntropy(MolarEntropyInterface):
         return f"{self.long_name} in liquid bismuth"
 
 
-class GibbsFreeEnergy(PropertyInterface):
+class GibbsFreeEnergy(GibbsFreeEnergyInterface):
     """
     Liquid bismuth *Gibbs free energy variation* property class.
     """
@@ -382,33 +382,12 @@ class GibbsFreeEnergy(PropertyInterface):
             - T * MolarEntropy().correlation(T, p)
 
     @property
-    def name(self) -> str:
-        """
-        str : Name of the property
-        """
-        return "G"
-
-    @property
-    def units(self) -> str:
-        """
-        str : Gibbs free energy unit
-        """
-        return "[J/mol]"
-
-    @property
     def range(self) -> List[float]:
         """
         List[float] : Temperature validity range of the Gibbs free energy
         variation correlation function
         """
         return [T_m0, T_b0]
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Gibbs free energy long name
-        """
-        return "Gibbs free energy variation"
 
     @property
     def description(self) -> str:
